vulkan: code cleanup

+++ vtest/vtest_vk.h
@@ -10,7 +10,7 @@ int vtest_vk_read_memory(uint32_t length_dw);
 int vtest_vk_write_memory(uint32_t length_dw);
 
 #define CHECK_IO_RESULT(Done, Expected)                                    \
-   if ((Done) < (Expected)) {                                              \
+   if ((Done) < (int)(Expected)) {                                         \
       fprintf(stderr, "%s: failed to write back the answer.\n", __func__); \
       RETURN(-1);                                                          \
    }
